# Backend/src/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SchooliAT API",
    "description": "School Management Application API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ReceiptStatus": {
        "type": "string",
        "enum": ["GENERATED", "SENT", "PAID", "CANCELLED"]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/schools": {
      "post": {
        "tags": ["Schools"],
        "summary": "Create a school",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["request"],
                "properties": {
                  "request": {
                    "type": "object",
                    "required": ["name", "email", "phone", "address", "code"],
                    "properties": {
                      "name": { "type": "string" },
                      "email": { "type": "string", "format": "email" },
                      "phone": { "type": "string" },
                      "address": {
                        "type": "array",
                        "items": { "type": "string" }
                      },
                      "code": { "type": "string" },
                      "gstNumber": { "type": "string" },
                      "principalName": { "type": "string" },
                      "principalEmail": { "type": "string" },
                      "principalPhone": { "type": "string" },
                      "establishedYear": { "type": "string" },
                      "boardAffiliation": { "type": "string" },
                      "studentStrength": { "type": "string" },
                      "certificateLink": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "School created" }
        }
      },
      "get": {
        "tags": ["Schools"],
        "summary": "Get all schools",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Schools fetched" }
        }
      }
    },
...
    "/receipts": {
      "post": {
        "tags": ["Receipts"],
        "summary": "Create a receipt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["request"],
                "properties": {
                  "request": {
                    "type": "object",
                    "required": [
                      "receiptNumber",
                      "schoolId",
                      "baseAmount",
                      "paymentMethod"
                    ],
                    "properties": {
                      "receiptNumber": { "type": "string" },
                      "schoolId": { "type": "string", "format": "uuid" },
                      "baseAmount": { "type": "string" },
                      "sgstPercent": { "type": "string" },
                      "cgstPercent": { "type": "string" },
                      "igstPercent": { "type": "string" },
                      "ugstPercent": { "type": "string" },
                      "description": { "type": "string" },
                      "paymentMethod": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Receipt created" }
        }
      },
      "get": {
        "tags": ["Receipts"],
        "summary": "Get all receipts",
        "parameters": [
          {
            "name": "schoolId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ReceiptStatus" }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Receipts fetched" }
        }
      }
    },
    "/receipts/{id}": {
      "get": {
        "tags": ["Receipts"],
        "summary": "Get a receipt by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Receipt fetched" }
        }
      },
      "put": {
        "tags": ["Receipts"],
        "summary": "Update a receipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "request": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "$ref": "#/components/schemas/ReceiptStatus"
                      },
                      "description": { "type": "string" },
                      "paymentMethod": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Receipt updated" }
        }
      },
      "delete": {
        "tags": ["Receipts"],
        "summary": "Delete a receipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Receipt deleted" }
        }
      }
    },
    "/receipts/{id}/generate": {
      "post": {
        "tags": ["Receipts"],
        "summary": "Generate receipt HTML for printing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "html": { "type": "string" },
                        "printUrl": { "type": "string" },
                        "receipt": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/statistics/schools": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get school statistics",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "School statistics fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "schools": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "name": { "type": "string" },
                              "code": { "type": "string" },
                              "totalStudents": { "type": "integer" },
                              "totalStaff": { "type": "integer" },
                              "teachers": { "type": "integer" },
                              "adminStaff": { "type": "integer" },
                              "status": { "type": "string" }
                            }
                          }
                        },
                        "totals": {
                          "type": "object",
                          "properties": {
                            "totalStudents": { "type": "integer" },
                            "totalStaff": { "type": "integer" },
                            "totalTeachers": { "type": "integer" },
                            "totalAdminStaff": { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}